Add bit-identical replay script with depth capture (#4)
* Add replay script for recorded Zarr demos

scripts/replay_demos_zarr.py replays demos collected by
record_demos_zarr.py against a freshly reset env, restoring only the
recorded object poses and table height while letting lighting, table
texture, and the third-person camera stay randomized.

Also realigns pick_bucket's proprio_keys to put bucket before
boxed_food, matching the env's _compute_observation dict order and the
ordering used by every other task config.

* Keep proprio state in float64 to make replay bit-identical

DexjocoObsAdapter was casting the flat proprio vector to float32 on the
way out, and water_plant's _compute_observation packed
table_delta_height as float32 inline. Both lost ~1e-8 of precision per
field at the record/replay boundary, which the MuJoCo contact solver
amplified to ~1e-2 after ~100 steps on contact-heavy tasks like
bimanual_microwave_cook. Promoting both spots to float64 makes record
and replay reproduce identical physical trajectories across visual
randomization seeds (max diff drops from 2.6e-2 to 0.0 across all 11
tasks over 120 steps).

* Move per-task restorers to dexjoco.tasks.state_restorers

The replay script shrinks from 567 to 349 lines and now imports
`restore_initial_state` / `has_restorer` from the new module. The
public entry point is unchanged; this is purely a relocation, plus
two fidelity tweaks:

- `_restore_click_mouse` now also updates `display_root.qpos[2]` to
  `_display_root_z0 + restored_delta_h`. The monitor is a free joint
  with gravity; reset() pins its z to the originally-sampled delta_h,
  so after the restorer overwrites delta_h the monitor would float or
  bury. Across visual seeds this leaked into the contact solver and
  compounded to ~1e-1 drift by the end of an episode.
- `_restore_click_mouse` also re-attaches the mousepad to the recorded
  mouse_xy at the correct table z (read from the table site rather
  than `_table_body_z0 + delta_h`), so the contact graph matches.

With these in place, record-vs-replay diff is 0.0 on all 11 tasks /
12 example demos across different visual seeds (full episode length).

* Seed the RNGs in bimanual_microwave_cook env

The other 10 envs seed `random` and `np.random` from the constructor's
`seed` argument; this one was missing those calls, so its reset()
sampling was sensitive to the numpy global RNG state of whatever ran
before it in the same process. Cross-process record/replay still
worked because the recorded state[0] pins everything the restorer
needs, but same-process tests had to manually re-seed numpy to get
deterministic comparisons.

* Capture depth videos alongside RGB in record and replay

Both scripts grow a --save_depth flag (off by default). When enabled,
each step also renders a depth_array per RGB camera, and a new helper
module dexjoco.data.depth_capture writes two artefacts per camera:

- <cam>_depth.npz: float32 depth in metres, shape (T, H, W); sky
  pixels are inf so consumers can mask them.
- <cam>_depth.mp4: gray-scale H.264 preview, normalised to the
  per-clip min/max of finite pixels.

mujoco's depth_array is the GL z-buffer in [0,1]; depth_capture
linearises it to metres using znear*extent and zfar*extent.

Verified bit-identical physics across visual seeds on all 11 tasks:
state diff = 0, wrist (hand-mounted) depth diff = 0, scene-camera
depth diff = ~3-5m (visual randomisation moves those cameras by
design).

* Default --extend_steps to 0 in replay

Now that record/replay are bit-identical, the recorded actions already
end on the succeed step and no padding is needed. The flag stays in
case someone replays older data that was clipped before success.

+++ dexjoco/dexjoco/tasks/obs_adapters.py
@@ -29,9 +29,9 @@ class DexjocoObsAdapter(gym.ObservationWrapper):
     def observation(self, obs):
         proprio_values = [np.asarray(obs["state"][key]).ravel() for key in self.proprio_keys]
         if proprio_values:
-            flat_state = np.concatenate(proprio_values, axis=0).astype(np.float32, copy=False)
+            flat_state = np.concatenate(proprio_values, axis=0).astype(np.float64, copy=False)
         else:
-            flat_state = np.zeros((0,), dtype=np.float32)
+            flat_state = np.zeros((0,), dtype=np.float64)
         return {"state": flat_state, **obs.get("images", {})}
 
     def reset(self, **kwargs):
